[#184] fix: resolution usecase 명명 수정
