+++ lib/presentation/late_writing/provider/late_writing_view_provider.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:wehavit/common/utils/utils.dart';
+import 'package:wehavit/common/utils/no_params.dart';
 import 'package:wehavit/domain/entities/confirm_post_entity/confirm_post_entity.dart';
 import 'package:wehavit/domain/entities/resolution_entity/resolution_entity.dart';
-import 'package:wehavit/domain/usecases/get_resolution_list_usecase.dart';
+import 'package:wehavit/domain/usecases/get_my_resolution_list_usecase.dart';
 import 'package:wehavit/domain/usecases/upload_confirm_post_usecase.dart';
 import 'package:wehavit/presentation/late_writing/model/late_writing_view_model.dart';
 
@@ -13,7 +13,7 @@ final lateWritingViewModelProvider =
   (ref) {
     final createPostUsecase = ref.watch(uploadConfirmPostUseCaseProvider);
     final getMyResolutionListUsecase =
-        ref.watch(getResolutionListByUserIdUsecaseProvider);
+        ref.watch(getMyResolutionListByUserIdUsecaseProvider);
 
     return LateWritingViewModelProvider(
       createPostUsecase,
@@ -23,9 +23,6 @@ final lateWritingViewModelProvider =
 );
 
 class LateWritingViewModelProvider extends StateNotifier<LateWritingViewModel> {
-  final UploadConfirmPostUseCase _createPostUseCase;
-  final GetResolutionListByUserIdUsecase _getMyResolutionListUsecase;
-
   LateWritingViewModelProvider(
     this._createPostUseCase,
     this._getMyResolutionListUsecase,
@@ -33,11 +30,12 @@ class LateWritingViewModelProvider extends StateNotifier<LateWritingViewModel> {
     // 팩토리 메서드 호출
     _initialize();
   }
+  final UploadConfirmPostUseCase _createPostUseCase;
+  final GetMyResolutionListByUserIdUsecase _getMyResolutionListUsecase;
 
   // 비동기적으로 초기화를 진행하는 팩토리 메서드
   Future<void> _initialize() async {
-    // TODO: 자신의 resolution list를 받아오는 코드 작성하기
-    state.resolutionList = _getMyResolutionListUsecase('hello');
+    state.resolutionList = _getMyResolutionListUsecase(NoParams());
   }
 
   Future<void> postCurrentConfirmPost() async {
