+++ lib/presentation/my_page/providers/my_page_resolution_list_provider.dart
@@ -36,7 +36,7 @@ class MyPageResolutionListProvider extends StateNotifier<
       getConfirmPostListForResolutionIdUsecase;
 
   Future<void> getMyActiveResolutionList() async {
-    final resolutionFetchResult = await getMyResolutionListUsecase('hello');
+    final resolutionFetchResult = await getMyResolutionListUsecase(NoParams());
 
     final List<ResolutionEntity> resolutionList = resolutionFetchResult.fold(
       (failure) => [],
